[DMS-975] Change DatabaseFingerprint.ResourceKeyCount from int to short to match SMALLINT DDL

+++ src/dms/core/EdFi.DataManagementService.Core.External/Backend/DatabaseFingerprint.cs
@@ -15,6 +15,6 @@ namespace EdFi.DataManagementService.Core.External.Backend;
 public sealed record DatabaseFingerprint(
     string ApiSchemaFormatVersion,
     string EffectiveSchemaHash,
-    int ResourceKeyCount,
+    short ResourceKeyCount,
     byte[] ResourceKeySeedHash
 );
